Support opening specs in adjacent panes
This is still kind of crappy because of how the settings work. Also
`left` doesn’t work too well because it doesn’t *create* a pane to the
left but instead uses the leftmost pane. This will suffice for now
though since the default of `right` seems like the most desirable
option anyway.